flag on the testclient to catch unhandle server exceptions
Following the same semantics on starlette, where, exceptions
are reraised to allow server process to catch them,
and show the proper stacktrace on logs, we need to provide
a flag on test client to catch them.

Anyway, seems like there is still not working the exception
propagation, I opened an issue on:

https://github.com/encode/starlette/issues/484

But right now, at least to be compact with starlette,
we need to provide this flag, like his own sync TestClient provides.

I know that with Quart nothing happens and this works as expected