Add tests for ASGI specification compliance
There are a number of tests that currently xfail in this suite.
The intention is to fix the issues one by one, or in cases where
the issue is a WONTFIX, change the test and/or indicate as such
in the xfail reason.

There is still a need for a couple more tests in the websocket
section, and probably could do with some more type-oriented
negative tests (e.g. testing what happens if you send a message
with the wrong type in the event values, like with 'status')