Disable lifespan protocol if apps do not support it
Fixes test_lifespan:test_lifespan_not_supported_is_allowed()

This fix was somewhat challenging to achieve without breaking
other things or backward compatibility issues.

In the end, I think the most elegant solution is to adopt
more Pythonic exceptions in the TestClient, as demonstrated in
the commit. However, since this is a less trivial change, I
have only used TestClientError in the lifespan related code,
 leaving all other code using raw Exceptions still.

I think TestClientError (or possibly even multiple exception classes
for different purposes) should actually be adopted throughout
all the code, albeit with the caveat that it might cause some issues
if anyone is doing really rigid exception assertions in their test
code. For most cases (including my own code that is already using
async_asgi_testclient) I believe the average
 'with pytest.raises(Exception)' assertions should still work,
 since TestClientError is a sub-class of Exception.

A side-benefit of the TestClientError is it would allow access to
the original Message structure, instead of having to decode a string
repr of the event message.